+++ platform-api/src/main/java/com/platform/entity/UserCouponVo.java
@@ -18,7 +18,7 @@ public class UserCouponVo implements Serializable {
     /**主键**/
     private Long id;
     /**优惠券Id**/
-    private Integer coupon_id;
+    private Long coupon_id;
     /**优惠券数量**/
     private String coupon_number;
     /**会员Id**/
