+++ platform-api/src/main/java/com/platform/entity/UserLevelEnum.java
@@ -16,30 +16,30 @@ public enum UserLevelEnum {
     /**
      * 普通会员
      **/
-    NORMAL(1, "普通会员"),
+    NORMAL(1L, "普通会员"),
     /**
      * 会员
      **/
-    VIP(2, "会员"),
+    VIP(2L, "会员"),
     /**
      * 创客
      **/
-    CHUANGKE(3, "创客"),
+    CHUANGKE(3L, "创客"),
     /**
      * 城市合伙人
      **/
-    HEHUOREN(4, "城市合伙人");
+    HEHUOREN(4L, "城市合伙人");
 
-    UserLevelEnum(Integer levelId, String levelName) {
+    UserLevelEnum(Long levelId, String levelName) {
         this.levelId = levelId;
         this.levelName = levelName;
     }
 
-    private Integer levelId;
+    private Long levelId;
 
     private String levelName;
 
-    public static Map<Integer, UserLevelEnum> LEVEL_MAP = new HashMap<>();
+    public static Map<Long, UserLevelEnum> LEVEL_MAP = new HashMap<>();
 
     static {
         Arrays.asList(UserLevelEnum.values()).forEach(e -> LEVEL_MAP.put(e.getLevelId(), e));
