+++ platform-api/src/main/java/com/platform/service/ApiKeywordsService.java
@@ -45,7 +45,7 @@ public class ApiKeywordsService {
     }
 
 
-    public void deleteBatch(Integer[] ids) {
+    public void deleteBatch(Long[] ids) {
         keywordsDao.deleteBatch(ids);
     }
